+++ Changelog.md
@@ -1,1 +1,7 @@
+# Changelog
+
+### Version 1.16.5 - 1.0.0
+ - add forgotten well structure
+ - add jungle temple structure
+ - add underground temple structure